feat: add the "ecdsa_example" macro branch
